factored communication out of ThingManager into own component
...in order to separate this concern from the rest of the diverse
duties of the ThingManager.
