+++ bundles/core/org.eclipse.smarthome.core.thing/src/main/java/org/eclipse/smarthome/core/thing/internal/ThingManager.java
@@ -38,12 +38,7 @@ import org.eclipse.smarthome.core.common.registry.ManagedProvider;
 import org.eclipse.smarthome.core.common.registry.Provider;
 import org.eclipse.smarthome.core.events.EventPublisher;
 import org.eclipse.smarthome.core.events.EventSubscriber;
-import org.eclipse.smarthome.core.items.Item;
-import org.eclipse.smarthome.core.items.ItemUtil;
 import org.eclipse.smarthome.core.items.events.AbstractItemEventSubscriber;
-import org.eclipse.smarthome.core.items.events.ItemCommandEvent;
-import org.eclipse.smarthome.core.items.events.ItemEventFactory;
-import org.eclipse.smarthome.core.items.events.ItemStateEvent;
 import org.eclipse.smarthome.core.service.ReadyMarker;
 import org.eclipse.smarthome.core.thing.Bridge;
 import org.eclipse.smarthome.core.thing.Channel;
@@ -63,7 +58,6 @@ import org.eclipse.smarthome.core.thing.binding.ThingHandlerFactory;
 import org.eclipse.smarthome.core.thing.binding.builder.ThingStatusInfoBuilder;
 import org.eclipse.smarthome.core.thing.events.ThingEventFactory;
 import org.eclipse.smarthome.core.thing.i18n.ThingStatusInfoI18nLocalizationService;
-import org.eclipse.smarthome.core.thing.link.ItemChannelLinkRegistry;
 import org.eclipse.smarthome.core.thing.type.ThingType;
 import org.eclipse.smarthome.core.thing.type.ThingTypeRegistry;
 import org.eclipse.smarthome.core.thing.util.ThingHandlerHelper;
@@ -117,7 +111,7 @@ public class ThingManager extends AbstractItemEventSubscriber implements ThingTr
 
     private EventPublisher eventPublisher;
 
-    private ItemChannelLinkRegistry itemChannelLinkRegistry;
+    private CommunicationManager communicationManager;
 
     private List<ThingHandlerFactory> thingHandlerFactories = new CopyOnWriteArrayList<>();
 
@@ -136,20 +130,17 @@ public class ThingManager extends AbstractItemEventSubscriber implements ThingTr
 
         @Override
         public void stateUpdated(ChannelUID channelUID, State state) {
-            Set<Item> items = itemChannelLinkRegistry.getLinkedItems(channelUID);
-            for (Item item : items) {
-                State acceptedState = ItemUtil.convertToAcceptedState(state, item);
-                eventPublisher
-                        .post(ItemEventFactory.createStateEvent(item.getName(), acceptedState, channelUID.toString()));
-            }
+            communicationManager.stateUpdated(channelUID, state);
         }
 
         @Override
         public void postCommand(ChannelUID channelUID, Command command) {
-            Set<String> items = itemChannelLinkRegistry.getLinkedItemNames(channelUID);
-            for (String item : items) {
-                eventPublisher.post(ItemEventFactory.createCommandEvent(item, command, channelUID.toString()));
-            }
+            communicationManager.postCommand(channelUID, command);
+        }
+
+        @Override
+        public void channelTriggered(Thing thing, ChannelUID channelUID, String event) {
+            communicationManager.channelTriggered(thing, channelUID, event);
         }
 
         @Override
@@ -253,11 +244,6 @@ public class ThingManager extends AbstractItemEventSubscriber implements ThingTr
         public void migrateThingType(final Thing thing, final ThingTypeUID thingTypeUID,
                 final Configuration configuration) {
             ThingManager.this.migrateThingType(thing, thingTypeUID, configuration);
-        }
-
-        @Override
-        public void channelTriggered(Thing thing, ChannelUID channelUID, String event) {
-            eventPublisher.post(ThingEventFactory.createTriggerEvent(event, channelUID));
         }
 
     };
@@ -366,109 +352,6 @@ public class ThingManager extends AbstractItemEventSubscriber implements ThingTr
                 throw new RuntimeException(message);
             }
         }, 0, TimeUnit.MILLISECONDS);
-    }
-
-    @Override
-    protected void receiveCommand(ItemCommandEvent commandEvent) {
-        String itemName = commandEvent.getItemName();
-        final Command command = commandEvent.getItemCommand();
-        Set<ChannelUID> boundChannels = this.itemChannelLinkRegistry.getBoundChannels(itemName);
-        for (final ChannelUID channelUID : boundChannels) {
-            // make sure a command event is not sent back to its source
-            if (!channelUID.toString().equals(commandEvent.getSource())) {
-                Thing thing = getThing(channelUID.getThingUID());
-                if (thing != null) {
-                    final ThingHandler handler = thing.getHandler();
-                    if (handler != null) {
-                        if (ThingHandlerHelper.isHandlerInitialized(thing)) {
-                            logger.debug("Delegating command '{}' for item '{}' to handler for channel '{}'", command,
-                                    itemName, channelUID);
-                            try {
-                                SafeMethodCaller.call(new SafeMethodCaller.ActionWithException<Void>() {
-                                    @Override
-                                    public Void call() throws Exception {
-                                        handler.handleCommand(channelUID, command);
-                                        return null;
-                                    }
-                                });
-                            } catch (TimeoutException ex) {
-                                logger.warn("Handler for thing '{}' takes more than {}ms for handling a command",
-                                        handler.getThing().getUID(), SafeMethodCaller.DEFAULT_TIMEOUT);
-                            } catch (Exception ex) {
-                                logger.error("Exception occurred while calling handler: {}", ex.getMessage(), ex);
-                            }
-                        } else {
-                            logger.debug("Not delegating command '{}' for item '{}' to handler for channel '{}', "
-                                    + "because handler is not initialized (thing must be in status UNKNOWN, ONLINE or OFFLINE).",
-                                    command, itemName, channelUID);
-                        }
-                    } else {
-                        logger.warn("Cannot delegate command '{}' for item '{}' to handler for channel '{}', "
-                                + "because no handler is assigned. Maybe the binding is not installed or not "
-                                + "propertly initialized.", command, itemName, channelUID);
-                    }
-                } else {
-                    logger.warn(
-                            "Cannot delegate command '{}' for item '{}' to handler for channel '{}', "
-                                    + "because no thing with the UID '{}' could be found.",
-                            command, itemName, channelUID, channelUID.getThingUID());
-                }
-            }
-        }
-    }
-
-    @Override
-    protected void receiveUpdate(ItemStateEvent updateEvent) {
-        String itemName = updateEvent.getItemName();
-        final State newState = updateEvent.getItemState();
-        Set<ChannelUID> boundChannels = this.itemChannelLinkRegistry.getBoundChannels(itemName);
-        for (final ChannelUID channelUID : boundChannels) {
-            // make sure an update event is not sent back to its source
-            if (!channelUID.toString().equals(updateEvent.getSource())) {
-                Thing thing = getThing(channelUID.getThingUID());
-                if (thing != null) {
-                    final ThingHandler handler = thing.getHandler();
-                    if (handler != null) {
-                        if (ThingHandlerHelper.isHandlerInitialized(thing)) {
-                            logger.debug("Delegating update '{}' for item '{}' to handler for channel '{}'", newState,
-                                    itemName, channelUID);
-                            try {
-                                SafeMethodCaller.call(new SafeMethodCaller.ActionWithException<Void>() {
-                                    @Override
-                                    public Void call() throws Exception {
-                                        if (newState != null) {
-                                            handler.handleUpdate(channelUID, newState);
-                                        } else {
-                                            throw new IllegalStateException(
-                                                    "Trying to set state to null on channel " + channelUID);
-                                        }
-                                        return null;
-                                    }
-                                });
-                            } catch (TimeoutException ex) {
-                                logger.warn("Handler for thing {} takes more than {}ms for handling an update",
-                                        handler.getThing().getUID(), SafeMethodCaller.DEFAULT_TIMEOUT);
-                            } catch (Exception ex) {
-                                logger.error("Exception occurred while calling handler: {}", ex.getMessage(), ex);
-                            }
-                        } else {
-                            logger.debug("Not delegating update '{}' for item '{}' to handler for channel '{}', "
-                                    + "because handler is not initialized (thing must be in status UNKNOWN, ONLINE or OFFLINE).",
-                                    newState, itemName, channelUID);
-                        }
-                    } else {
-                        logger.warn("Cannot delegate update '{}' for item '{}' to handler for channel '{}', "
-                                + "because no handler is assigned. Maybe the binding is not installed or not "
-                                + "propertly initialized.", newState, itemName, channelUID);
-                    }
-                } else {
-                    logger.warn(
-                            "Cannot delegate update '{}' for item '{}' to handler for channel '{}', "
-                                    + "because no thing with the UID '{}' could be found.",
-                            newState, itemName, channelUID, channelUID.getThingUID());
-                }
-            }
-        }
     }
 
     @Override
@@ -1118,21 +1001,12 @@ public class ThingManager extends AbstractItemEventSubscriber implements ThingTr
     }
 
     @Reference
-    protected void setItemChannelLinkRegistry(ItemChannelLinkRegistry itemChannelLinkRegistry) {
-        this.itemChannelLinkRegistry = itemChannelLinkRegistry;
-    }
-
-    @Reference
     protected void setThingRegistry(ThingRegistry thingRegistry) {
         this.thingRegistry = (ThingRegistryImpl) thingRegistry;
     }
 
     protected void unsetEventPublisher(EventPublisher eventPublisher) {
         this.eventPublisher = null;
-    }
-
-    protected void unsetItemChannelLinkRegistry(ItemChannelLinkRegistry itemChannelLinkRegistry) {
-        this.itemChannelLinkRegistry = null;
     }
 
     protected void unsetThingRegistry(ThingRegistry thingRegistry) {
@@ -1192,6 +1066,15 @@ public class ThingManager extends AbstractItemEventSubscriber implements ThingTr
     protected void unsetThingStatusInfoI18nLocalizationService(
             ThingStatusInfoI18nLocalizationService thingStatusInfoI18nLocalizationService) {
         this.thingStatusInfoI18nLocalizationService = null;
+    }
+
+    @Reference
+    protected void setInboundCommunication(CommunicationManager communicationManager) {
+        this.communicationManager = communicationManager;
+    }
+
+    protected void unsetInboundCommunication(CommunicationManager communicationManager) {
+        this.communicationManager = null;
     }
 
 }
